Pruebas

+++ Arrays/arrays2.c
@@ -10,7 +10,7 @@
 
 int main(){
 
-float [5];
+int [5];
 int i;
 
 for (i = 0; i < 5; i++){
